refactoring: MainController 분리
1. Table~~ 클래스들 IllegalCar~~ 클래스로 Rename

Resolves:
Ref:
Related to:
