+++ demo/src/main/java/com/example/demo/controller/MainController.java
@@ -1,31 +1,21 @@
 package com.example.demo.controller;
 
 import com.example.demo.DTO.UserRegistrationDto;
-import com.example.demo.Entity.Member;
 import com.example.demo.DTO.NotificationCarNumberDTO;
-import com.example.demo.DTO.carNumber;
-import com.example.demo.s3.S3Service;
 import com.example.demo.service.*;
-import lombok.AllArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.*;
 
 @Slf4j
 @Controller
 @RequiredArgsConstructor
 public class MainController {
-    private TableServiceInterface tableServiceInterface;
+    private IllegalCarServiceInterface illegalCarServiceInterface;
     private SmsService smsService;
 
     @GetMapping("mainPage/index")
@@ -57,7 +47,7 @@ public class MainController {
         if (notificationCarNumberDTO == null) {
             return "notification/notificationService";
         }
-        Optional<NotificationCarNumberDTO> DTOforGetEnteringTime = tableServiceInterface.isExist(notificationCarNumberDTO.getCarN());
+        Optional<NotificationCarNumberDTO> DTOforGetEnteringTime = illegalCarServiceInterface.isExist(notificationCarNumberDTO.getCarN());
         model.addAttribute("enteringTime", DTOforGetEnteringTime.get().getTimestamp());
 
         return "notification/current";
