+++ demo/src/main/java/com/example/demo/controller/NotificationController.java
@@ -3,7 +3,7 @@ package com.example.demo.controller;
 import com.example.demo.DTO.NotificationCarNumberDTO;
 import com.example.demo.service.NotificationSessionConst;
 import com.example.demo.service.Notification_Thread;
-import com.example.demo.service.TableServiceInterface;
+import com.example.demo.service.IllegalCarServiceInterface;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -25,12 +25,12 @@ import java.util.HashMap;
 @RequiredArgsConstructor
 public class NotificationController {
 
-    private final TableServiceInterface tableServiceInterface;
+    private final IllegalCarServiceInterface illegalCarServiceInterface;
     @PostMapping("notification/exittimeupdate")
     public ResponseEntity<?> carUpdateExitTime(@RequestBody HashMap<String, Object> map) throws ParseException {
         //차량 출차 시 출차 시간 업데이트.
         try {
-            tableServiceInterface.updateCurrentCarExitTime(map);
+            illegalCarServiceInterface.updateCurrentCarExitTime(map);
             return ResponseEntity.ok("출차 입력 성공");
         } catch (ParseException e) {
             // ParseException이 발생했을 때 실패 응답 반환
@@ -45,9 +45,15 @@ public class NotificationController {
         NotificationCarNumberDTO notificationCarNumberDTO = new NotificationCarNumberDTO();
         notificationCarNumberDTO.setCarN((String) map.get("carNumber"));
         //알림서비스 스레드 동작
-        Notification_Thread notification_thread = new Notification_Thread(smsService, tableServiceInterface, (String) map.get("carNumber"));
+        Notification_Thread notification_thread = new Notification_Thread(smsService, illegalCarServiceInterface, (String) map.get("carNumber"));
         notification_thread.start();
         //알림서비스 스레드 동작
+
+        /**
+         * 1. 맵을 넘겨주고 비동기 메소드 동작
+         * 2. 알림 서비스 등록차량 or 미등록차량 판별
+         * 3. 판별 결과에 따라 새로 등록할지 아니면 기존 등록 차량 정보를 업데이트할지 선택
+         */
 
         // 날짜처리코드
         String timeStr = (String) map.get("EnterDate");
@@ -62,13 +68,13 @@ public class NotificationController {
 
         notificationCarNumberDTO.setTimestamp(newTime);
 
-        if (!tableServiceInterface.isExist((String) map.get("carNumber")).isPresent()) {
+        if (!illegalCarServiceInterface.isExist((String) map.get("carNumber")).isPresent()) {
             // 알림 서비스 등록 안된 차량이면 일단 db에 전화번호 없이 신규등록
-            tableServiceInterface.NotificationCarRegister(notificationCarNumberDTO);
+            illegalCarServiceInterface.NotificationCarRegister(notificationCarNumberDTO);
         } else {
             // 알림 서비스 등록 경력 있는 차량이면 (or db에 이미 존재하는 차량번호) 입차시간 업데이트 해주고 출차시간 리셋
-            tableServiceInterface.updateEnteringTime((String)map.get("carNumber"), timestamp);
-            tableServiceInterface.resetNewCarExitTime((String)map.get("carNumber"));
+            illegalCarServiceInterface.updateEnteringTime((String)map.get("carNumber"), timestamp);
+            illegalCarServiceInterface.resetNewCarExitTime((String)map.get("carNumber"));
         }
     }
     @PostMapping("notification/notificationService")
@@ -78,13 +84,13 @@ public class NotificationController {
         String inputCarNumber = notificationCarNumberDTO.getCarN();
         String inputPhoneNumber = notificationCarNumberDTO.getPhoneNumber();
 
-        if (!tableServiceInterface.isExist(inputCarNumber).isPresent()) {
+        if (!illegalCarServiceInterface.isExist(inputCarNumber).isPresent()) {
             model.addAttribute("msg", "해당번호로 입차한 차량이 없습니다.");
             model.addAttribute("url", "notificationService");
             return "notification/messageRedirect";
         }
-        if (tableServiceInterface.isExistPhoneNumber(inputCarNumber) == null) {
-            tableServiceInterface.updatePhoneNumber(inputCarNumber, inputPhoneNumber);
+        if (illegalCarServiceInterface.isExistPhoneNumber(inputCarNumber) == null) {
+            illegalCarServiceInterface.updatePhoneNumber(inputCarNumber, inputPhoneNumber);
             model.addAttribute("msg", "주차 알림 서비스 신규등록이 완료되었습니다.");
             model.addAttribute("url", "current");
         } else {
@@ -93,7 +99,7 @@ public class NotificationController {
         }
 
         HttpSession session = request.getSession();
-        session.setAttribute(NotificationSessionConst.NOTIFY_CAR, tableServiceInterface.isExist(inputCarNumber).get());
+        session.setAttribute(NotificationSessionConst.NOTIFY_CAR, illegalCarServiceInterface.isExist(inputCarNumber).get());
 
         return "notification/messageRedirect";
     }
