--- demo/src/main/java/com/example/demo/service/TableServiceInterface.java
@@ -1,35 +1,0 @@
-package com.example.demo.service;
-
-import com.example.demo.DTO.NotificationCarNumberDTO;
-import com.example.demo.DTO.carNumber;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.sql.Timestamp;
-import java.text.ParseException;
-import java.util.*;
-
-public interface TableServiceInterface {
-    List<carNumber> getAll();
-
-    void illegalCarRegister(HashMap<String, Object> map) throws ParseException;
-
-    void illegalCarRemove(Long id);
-
-    void NotificationCarRegister(NotificationCarNumberDTO notificationCarNumberDTO);
-
-    Optional<NotificationCarNumberDTO> isExist(String carNumber);
-
-    String isExistPhoneNumber(String carNumber);
-
-    List<Long> getMonthData();
-    void updatePhoneNumber(String carNumber, String phoneNumber);
-    void updateEnteringTime(String carNumber, java.sql.Timestamp timestamp);
-
-    void resetNewCarExitTime(String carNumber);
-
-    void updateCurrentCarExitTime(HashMap<String, Object> map) throws ParseException;
-
-    Timestamp getEnteringCarTimestamp(String carNumber);
-
-    Boolean isOverTIme(String carNumber);
-}
